pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chargectl"
version = "0.1.0"
description = "Lightweight EV charger controller for Tesla Wall Connector Gen 2 via RS-485"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "paho-mqtt>=2.0",
    "pyserial>=3.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
chargectl = "chargectl.__main__:main"
